SphinxFullscreenImageView for the gesture enabled imageView
