+++ sphinx/screens/chats/chat-common/chat-common/src/main/res/layout/layout_attachment_fullscreen.xml
@@ -9,7 +9,7 @@
     android:visibility="gone"
     tools:visibility="visible">
 
-    <androidx.appcompat.widget.AppCompatImageView
+    <chat.sphinx.chat_common.ui.SphinxFullscreenImageView
         android:id="@+id/image_view_attachment_fullscreen"
         android:layout_width="match_parent"
         android:layout_height="0dp"
